chore: move libs only used in browser out of common
