+++ spec-main/internal-spec.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai';
 
 describe('feature-string parsing', () => {
   it('is indifferent to whitespace around keys and values', () => {
-    const { parseCommaSeparatedKeyValue } = require('../lib/common/parse-features-string');
+    const { parseCommaSeparatedKeyValue } = require('../lib/browser/parse-features-string');
     const checkParse = (string: string, parsed: Record<string, string | boolean>) => {
       const features = parseCommaSeparatedKeyValue(string);
       expect(features).to.deep.equal(parsed);
